Fix parenthesis highlighting
The original molokai highlighting made it look like the cursor was
jumping to the far bracket which was very disconcerting.

There is a discussion and example on the pull here - https://github.com/tomasr/molokai/pull/44

In that pull the colors are simply reversed, I've tweaked them slightly
so that the far cursor is a bit easier to spot by having a lightened
background.